Fix the types in the task scheduler

+++ src/threading/TaskScheduler.hpp
@@ -280,7 +280,7 @@ namespace threading {
         std::atomic<bool> started{false};
 
         /// A map of group ids to the number of active tasks currently running in that group
-        std::map<NUClear::id_t, size_t> groups;
+        std::map<NUClear::id_t, int> groups;
         /// Mutex for the group map
         std::mutex group_mutex;
 
@@ -289,7 +289,7 @@ namespace threading {
         /// Global idle tasks to be executed when no other tasks are running
         std::map<NUClear::id_t, std::function<void()>> idle_tasks;
         /// The total number of threads that have runnable tasks
-        std::atomic<size_t> global_runnable_tasks{0};
+        std::atomic<int> global_runnable_tasks{0};
 
         /// A map of pool descriptor ids to pool descriptors
         std::map<NUClear::id_t, std::shared_ptr<PoolQueue>> pool_queues;
